+++ examples/redis.rs
@@ -7,7 +7,7 @@ Cleanup the redis docker container:
     docker rm -f kash-redis-example
  */
 
-use kash::io_kash;
+use kash::{io_kash, RedisCacheError};
 use std::io;
 use std::io::Write;
 use std::time::Duration;
@@ -19,14 +19,15 @@ enum ExampleError {
     RedisError(String),
 }
 
+impl From<RedisCacheError> for ExampleError {
+    fn from(e: RedisCacheError) -> Self {
+        ExampleError::RedisError(format!("{:?}", e))
+    }
+}
+
 // When the macro constructs your RedisCache instance, the connection string
 // will be pulled from the env var: `KASH_REDIS_CONNECTION_STRING`;
-#[io_kash(
-    redis,
-    ttl = 30,
-    cache_prefix_block = r#"{ "cache-redis-example-1" }"#,
-    map_error = r#"|e| ExampleError::RedisError(format!("{:?}", e))"#
-)]
+#[io_kash(redis, ttl = "30", cache_prefix_block = r#"{ "cache-redis-example-1" }"#)]
 fn kash_sleep_secs(secs: u64) -> Result<(), ExampleError> {
     std::thread::sleep(Duration::from_secs(secs));
     Ok(())
@@ -34,11 +35,7 @@ fn kash_sleep_secs(secs: u64) -> Result<(), ExampleError> {
 
 // If not `cache_prefix_block` is specified, then the function name
 // is used to create a prefix for cache keys used by this function
-#[io_kash(
-    redis,
-    ttl = 30,
-    map_error = r#"|e| ExampleError::RedisError(format!("{:?}", e))"#
-)]
+#[io_kash(redis, ttl = "30")]
 fn kash_sleep_secs_example_2(secs: u64) -> Result<(), ExampleError> {
     std::thread::sleep(Duration::from_secs(secs));
     Ok(())