+++ src/lib.rs
@@ -67,7 +67,7 @@ fn keyed(a: &str, b: &str) -> usize {
 ----
 
 ```rust
-use kash::io_kash;
+use kash::{io_kash, RedisCacheError};
 use kash::AsyncRedisCache;
 use thiserror::Error;
 
@@ -77,15 +77,18 @@ enum ExampleError {
     RedisError(String),
 }
 
+impl From<RedisCacheError> for ExampleError {
+    fn from(e: RedisCacheError) -> Self {
+        ExampleError::RedisError(format!("{:?}", e))
+    }
+}
+
 /// Cache the results of an async function in redis. Cache
 /// keys will be prefixed with `cache_redis_prefix`.
 /// A `map_error` closure must be specified to convert any
 /// redis cache errors into the same type of error returned
 /// by your function. All `io_kash` functions must return `Result`s.
-#[io_kash(
-    map_error = r##"|e| ExampleError::RedisError(format!("{:?}", e))"##,
-    redis,
-)]
+#[io_kash(redis)]
 async fn async_kash_sleep_secs(secs: u64) -> Result<String, ExampleError> {
     std::thread::sleep(std::time::Duration::from_secs(secs));
     Ok(secs.to_string())
@@ -95,7 +98,7 @@ async fn async_kash_sleep_secs(secs: u64) -> Result<String, ExampleError> {
 ----
 
 ```rust
-use kash::io_kash;
+use kash::{io_kash, DiskCacheError};
 use kash::DiskCache;
 use thiserror::Error;
 
@@ -105,16 +108,19 @@ enum ExampleError {
     DiskError(String),
 }
 
+impl From<DiskCacheError> for ExampleError {
+    fn from(e: DiskCacheError) -> Self {
+        ExampleError::DiskError(format!("{:?}", e))
+    }
+}
+
 /// Cache the results of a function on disk.
 /// Cache files will be stored under the system cache dir
 /// unless otherwise specified with `disk_dir` or the `create` argument.
 /// A `map_error` closure must be specified to convert any
 /// disk cache errors into the same type of error returned
 /// by your function. All `io_kash` functions must return `Result`s.
-#[io_kash(
-    map_error = r##"|e| ExampleError::DiskError(format!("{:?}", e))"##,
-    disk
-)]
+#[io_kash(disk)]
 fn kash_sleep_secs(secs: u64) -> Result<String, ExampleError> {
     std::thread::sleep(std::time::Duration::from_secs(secs));
     Ok(secs.to_string())