+++ tests/disk.rs
@@ -1,6 +1,6 @@
 #![cfg(feature = "disk_store")]
 
-use kash::io_kash;
+use kash::{io_kash, DiskCacheError};
 use thiserror::Error;
 
 #[derive(Error, Debug, PartialEq, Clone)]
@@ -11,11 +11,13 @@ enum TestError {
     Count(u32),
 }
 
-#[io_kash(
-    disk,
-    ttl = 1,
-    map_error = r##"|e| TestError::DiskError(format!("{:?}", e))"##
-)]
+impl From<DiskCacheError> for TestError {
+    fn from(e: DiskCacheError) -> Self {
+        TestError::DiskError(format!("{:?}", e))
+    }
+}
+
+#[io_kash(disk, ttl = "1")]
 fn kash_disk(n: u32) -> Result<u32, TestError> {
     if n < 5 {
         Ok(n)
@@ -32,11 +34,7 @@ fn test_kash_disk() {
     assert_eq!(kash_disk(6), Err(TestError::Count(6)));
 }
 
-#[io_kash(
-    map_error = r##"|e| TestError::DiskError(format!("{:?}", e))"##,
-    ttl = "1",
-    disk
-)]
+#[io_kash(ttl = "1", disk)]
 fn kash_disk_cache_create(n: u32) -> Result<u32, TestError> {
     if n < 5 {
         Ok(n)
@@ -58,8 +56,7 @@ fn test_kash_disk_cache_create() {
 /// There are no simple tests to test this here
 #[io_kash(
     disk,
-    map_error = r##"|e| TestError::DiskError(format!("{:?}", e))"##,
-    connection_config = r##"sled::Config::new().flush_every_ms(None)"##
+    connection_config = r#"sled::Config::new().flush_every_ms(None)"#
 )]
 fn kash_disk_connection_config(n: u32) -> Result<u32, TestError> {
     if n < 5 {
@@ -71,11 +68,7 @@ fn kash_disk_connection_config(n: u32) -> Result<u32, TestError> {
 
 /// Just calling the macro with sync_to_disk_on_cache_change to test it doesn't break with an expected value
 /// There are no simple tests to test this here
-#[io_kash(
-    disk,
-    map_error = r##"|e| TestError::DiskError(format!("{:?}", e))"##,
-    sync_to_disk_on_cache_change
-)]
+#[io_kash(disk, sync_to_disk_on_cache_change)]
 fn kash_disk_sync_to_disk_on_cache_change(n: u32) -> Result<u32, TestError> {
     if n < 5 {
         Ok(n)
@@ -88,7 +81,7 @@ fn kash_disk_sync_to_disk_on_cache_change(n: u32) -> Result<u32, TestError> {
 mod async_test {
     use super::*;
 
-    #[io_kash(disk, map_error = r##"|e| TestError::DiskError(format!("{:?}", e))"##)]
+    #[io_kash(disk)]
     async fn async_kash_disk(n: u32) -> Result<u32, TestError> {
         if n < 5 {
             Ok(n)