+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2024-xx-yy
+
+### Added
+
+### Changed
+
+- Type of `ttl` in `io_kash` is string. So you can pass functions or consts. 
+
+### Removed
+
+
+## [0.1.2] - 2024-10-14
+
+### Removed
+
+- Remove `map_error` param. Instead, you can impl `From<DiskCacheError>` or `From<RedisCacheError>` for your result.
+
 ## [0.1.1] - 2024-10-14
 
 ### Removed